Refactor common cli and utils classes
