+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/common/cli/CliManagerFactory.java
@@ -12,8 +12,9 @@
  * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
  * #L%
  */
-package com.vmware.pscoe.iac.artifact.cli;
+package com.vmware.pscoe.iac.artifact.common.cli;
 
+import com.vmware.pscoe.iac.artifact.aria.operations.cli.CliManagerVrops;
 import com.vmware.pscoe.iac.artifact.aria.operations.configuration.ConfigurationVrops;
 
 public final class CliManagerFactory {
