+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/common/utils/ZipUtilities.java
@@ -12,7 +12,7 @@
  * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
  * #L%
  */
-package com.vmware.pscoe.iac.artifact.cli;
+package com.vmware.pscoe.iac.artifact.common.utils;
 
 import java.io.File;
 import java.io.FileInputStream;
